requirements.txt
yfinance>=0.2.40
pandas>=2.2
numpy>=1.26
streamlit>=1.35
plotly>=5.22
pyyaml>=6.0
